+++ android/src/main/kotlin/com/oohyugi/sms_otp_auto_verify/SmsBroadcastReceiver.kt
@@ -3,6 +3,7 @@ package com.oohyugi.sms_otp_auto_verify
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
+import android.util.Log
 import com.google.android.gms.auth.api.phone.SmsRetriever
 import com.google.android.gms.common.api.CommonStatusCodes
 import com.google.android.gms.common.api.Status