+++ lib/sms_otp_auto_verify.dart
@@ -1,4 +1,4 @@
 library sms_otp_auto_verify;
 
 export 'package:sms_otp_auto_verify/src/text_field_pin.dart';
-export 'package:sms_otp_auto_verify/src/sms_retrieved.dart';
+export 'package:sms_otp_auto_verify/src/sms_verification.dart';